force output files to use UTF-8 encoding
Pathlib write_text by default with encoding=None does not use UTF-8 on
Windows, which can lead to a problem when feature names contain certain
characters. This commit ensures that this is not a problem. Also, the
loading from json of certain classes is bugged due to the unreliability
of jsonpickle. One of these issues is fixed here as well.